pyproject.toml
[project]
name = "ventureoracle"
version = "0.1.0"
description = "Personal content engine, idea discovery, and prediction platform"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "click>=8.1",
    "rich>=13.0",
    "httpx>=0.27",
    "atoma>=0.0.17",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "schedule>=1.2",
    "fastapi>=0.111",
    "uvicorn>=0.30",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.5",
]

[project.scripts]
ventureoracle = "ventureoracle.cli:cli"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
